+++ src/main/java/vendingmachine/SystemMessage/NoticeMessage.java
@@ -1,13 +1,13 @@
 package vendingmachine.SystemMessage;
 
 public class NoticeMessage {
-    public static final String CHANGE_MESSAGE="잔돈";
-    public static final String ASK_DRINK_NAME_MESSAGE="구매할 상품명을 입력해주세요.";
-    public static final String BALANCE_COIN_MESSAGE="자판기가 보유한 동전";
-    public static final String ASK_MONEY_INPUT_MESSAGE="투입 금액을 입력해 주세요.";
-    public static final String ASK_DRINK_LIST_MESSAGE="상품명과 가격, 수량을 입력해 주세요.";
-    public static final String ASK_BALANCE_INPUT_MESSAGE="자판기가 보유하고 있는 금액을 입력해 주세요.";
-    public static final String INPUT_MONEY_MESSAGE="투입 금액: ";
-    public static final String WON_MESSAGE="원";
-    public static final String EA_MESSAGE="개";
+    public static final String CHANGE_MESSAGE = "잔돈";
+    public static final String ASK_DRINK_NAME_MESSAGE = "구매할 상품명을 입력해주세요.";
+    public static final String BALANCE_COIN_MESSAGE = "자판기가 보유한 동전";
+    public static final String ASK_MONEY_INPUT_MESSAGE = "투입 금액을 입력해 주세요.";
+    public static final String ASK_DRINK_LIST_MESSAGE = "상품명과 가격, 수량을 입력해 주세요.";
+    public static final String ASK_BALANCE_INPUT_MESSAGE = "자판기가 보유하고 있는 금액을 입력해 주세요.";
+    public static final String INPUT_MONEY_MESSAGE = "투입 금액: ";
+    public static final String WON_MESSAGE = "원";
+    public static final String EA_MESSAGE = "개";
 }
